+++ test/unit/test_arguments.py
@@ -48,8 +48,8 @@ def test_simple_dict():
           output='test_output', overwrite_output=False,
           source_vocab=None, target_vocab=None, source_factor_vocabs=[], target_factor_vocabs=[],
           shared_vocab=False, num_words=(0, 0),
-          word_min_count=(1, 1), pad_vocab_to_multiple_of=None,
-          no_bucketing=False, bucket_width=8, bucket_scaling=False, no_bucket_scaling=None, max_seq_len=(95, 95),
+          word_min_count=(1, 1), pad_vocab_to_multiple_of=8,
+          no_bucketing=False, bucket_width=8, bucket_scaling=False, max_seq_len=(95, 95),
           monitor_pattern=None, monitor_stat_func='mx_default')),
 
     # short parameters
@@ -68,8 +68,8 @@ def test_simple_dict():
           output='test_output', overwrite_output=False,
           source_vocab=None, target_vocab=None, source_factor_vocabs=[], target_factor_vocabs=[],
           shared_vocab=False, num_words=(0, 0),
-          word_min_count=(1, 1), pad_vocab_to_multiple_of=None,
-          no_bucketing=False, bucket_width=8, bucket_scaling=False, no_bucket_scaling=None, max_seq_len=(95, 95),
+          word_min_count=(1, 1), pad_vocab_to_multiple_of=8,
+          no_bucketing=False, bucket_width=8, bucket_scaling=False, max_seq_len=(95, 95),
           monitor_pattern=None, monitor_stat_func='mx_default'))
 ])
 def test_io_args(test_params, expected_params):
@@ -90,14 +90,12 @@ def test_logging_args(test_params, expected_params):
 @pytest.mark.parametrize("test_params, expected_params", [
     ('', dict(device_ids=[-1],
               use_cpu=False,
-              omp_num_threads=None,
               env=None,
               disable_device_locking=False,
               lock_dir='/tmp')),
     ('--device-ids 1 2 3 --use-cpu --disable-device-locking --lock-dir test_dir',
      dict(device_ids=[1, 2, 3],
           use_cpu=True,
-          omp_num_threads=None,
           env=None,
           disable_device_locking=True,
           lock_dir='test_dir'))
@@ -131,8 +129,7 @@ def test_device_args(test_params, expected_params):
               encoder=C.TRANSFORMER_TYPE,
               decoder=C.TRANSFORMER_TYPE,
               dtype='float32',
-              amp=False,
-              amp_scale_interval=2000))
+              amp=False))
 ])
 def test_model_parameters(test_params, expected_params):
     _test_args(test_params, expected_params, arguments.add_model_parameters)
@@ -182,7 +179,6 @@ def test_inference_args(test_params, expected_params):
     ('', dict(batch_size=4096,
               batch_type='word',
               batch_sentences_multiple_of=8,
-              round_batch_sizes_to_multiple_of=None,
               loss='cross-entropy-without-softmax-output',
               label_smoothing=0.1,
               length_task=None,
@@ -290,11 +286,10 @@ def test_tutorial_averaging_args(test_params, expected_params, expected_params_p
           shared_vocab=False,
           num_words=(0, 0),
           word_min_count=(1, 1),
-          pad_vocab_to_multiple_of=None,
+          pad_vocab_to_multiple_of=8,
           no_bucketing=False,
           bucket_width=8,
           bucket_scaling=False,
-          no_bucket_scaling=None,
           max_seq_len=(95, 95),
           min_num_shards=1,
           num_samples_per_shard=10000000,
@@ -327,11 +322,10 @@ def test_tutorial_prepare_data_cli_args(test_params, expected_params):
           shared_vocab=False,
           num_words=(0, 0),
           word_min_count=(1, 1),
-          pad_vocab_to_multiple_of=None,
+          pad_vocab_to_multiple_of=8,
           no_bucketing=False,
           bucket_width=8,
           bucket_scaling=False,
-          no_bucket_scaling=None,
           max_seq_len=(95, 95),
           min_num_shards=1,
           num_samples_per_shard=10000000,